+++ lib/sindpad_design_system.dart
@@ -24,5 +24,8 @@ export 'components/auth/login/app_login_config.dart';
 export 'components/auth/login/app_login_form.dart';
 export 'components/auth/login/app_login_models.dart';
 
+// Feedback Components
+export 'components/feedback/sign_in_snack_bar.dart';
+
 // Third-party Icons
 export 'package:font_awesome_flutter/font_awesome_flutter.dart';